+++ docs/getting_started/jaeger_example.py
@@ -14,17 +14,18 @@
 
 # jaeger_example.py
 from opentelemetry import trace
-from opentelemetry.exporter import jaeger
+from opentelemetry.exporter.jaeger.thrift import JaegerExporter
+from opentelemetry.sdk.resources import SERVICE_NAME, Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchSpanProcessor
 
-trace.set_tracer_provider(TracerProvider())
-
-jaeger_exporter = jaeger.JaegerExporter(
-    service_name="my-helloworld-service",
-    agent_host_name="localhost",
-    agent_port=6831,
+trace.set_tracer_provider(
+    TracerProvider(
+        resource=Resource.create({SERVICE_NAME: "my-helloworld-service"})
+    )
 )
+
+jaeger_exporter = JaegerExporter(agent_host_name="localhost", agent_port=6831,)
 
 trace.get_tracer_provider().add_span_processor(
     BatchSpanProcessor(jaeger_exporter)