+++ exporter/opentelemetry-exporter-otlp-proto-grpc/tests/test_otlp_exporter_mixin.py
@@ -1,0 +1,47 @@
+# Copyright The OpenTelemetry Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from unittest import TestCase
+from unittest.mock import patch
+
+from grpc import Compression
+
+from opentelemetry.exporter.otlp.proto.grpc.exporter import (
+    InvalidCompressionValueException,
+    environ_to_compression,
+)
+
+
+class TestOTLPExporterMixin(TestCase):
+    def test_environ_to_compression(self):
+        with patch.dict(
+            "os.environ",
+            {
+                "test_gzip": "gzip",
+                "test_gzip_caseinsensitive_with_whitespace": " GzIp ",
+                "test_invalid": "some invalid compression",
+            },
+        ):
+            self.assertEqual(
+                environ_to_compression("test_gzip"), Compression.Gzip
+            )
+            self.assertEqual(
+                environ_to_compression(
+                    "test_gzip_caseinsensitive_with_whitespace"
+                ),
+                Compression.Gzip,
+            )
+            self.assertIsNone(environ_to_compression("missing_key"),)
+            with self.assertRaises(InvalidCompressionValueException):
+                environ_to_compression("test_invalid")