+++ exporter/opentelemetry-exporter-zipkin/src/opentelemetry/exporter/zipkin/__init__.py
@@ -27,8 +27,6 @@ v2 json, v2 protobuf).
 .. _OpenTelemetry: https://github.com/open-telemetry/opentelemetry-python/
 .. _Specification: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/sdk-environment-variables.md#zipkin-exporter
 
-.. envvar:: OTEL_EXPORTER_ZIPKIN_ENDPOINT
-
 .. code:: python
 
     from opentelemetry import trace
@@ -61,9 +59,7 @@ v2 json, v2 protobuf).
 
 The exporter supports the following environment variable for configuration:
 
-:envvar:`OTEL_EXPORTER_ZIPKIN_ENDPOINT`: zipkin collector endpoint to which the
-exporter will send data. This may include a path (e.g.
-http://example.com:9411/api/v2/spans).
+- :envvar:`OTEL_EXPORTER_ZIPKIN_ENDPOINT`
 
 API
 ---
@@ -75,7 +71,11 @@ from typing import Optional, Sequence
 
 import requests
 
-from opentelemetry.exporter.zipkin.encoder import Encoder, Protocol
+from opentelemetry.exporter.zipkin.encoder import (
+    DEFAULT_MAX_TAG_VALUE_LENGTH,
+    Encoder,
+    Protocol,
+)
 from opentelemetry.exporter.zipkin.encoder.v1.json import JsonV1Encoder
 from opentelemetry.exporter.zipkin.encoder.v2.json import JsonV2Encoder
 from opentelemetry.exporter.zipkin.encoder.v2.protobuf import ProtobufEncoder
@@ -83,6 +83,7 @@ from opentelemetry.exporter.zipkin.node_endpoint import IpInput, NodeEndpoint
 from opentelemetry.sdk.environment_variables import (
     OTEL_EXPORTER_ZIPKIN_ENDPOINT,
 )
+from opentelemetry.sdk.resources import SERVICE_NAME
 from opentelemetry.sdk.trace.export import SpanExporter, SpanExportResult
 from opentelemetry.trace import Span
 
@@ -120,6 +121,15 @@ class ZipkinExporter(SpanExporter):
             self.encoder = ProtobufEncoder(max_tag_value_length)
 
     def export(self, spans: Sequence[Span]) -> SpanExportResult:
+        # Populate service_name from first span
+        # We restrict any SpanProcessor to be only associated with a single
+        # TracerProvider, so it is safe to assume that all Spans in a single
+        # batch all originate from one TracerProvider (and in turn have all
+        # the same service.name)
+        if spans:
+            service_name = spans[0].resource.attributes.get(SERVICE_NAME)
+            if service_name:
+                self.local_node.service_name = service_name
         result = requests.post(
             url=self.endpoint,
             data=self.encoder.serialize(spans, self.local_node),